(PC-37958)[API] refactor: move `boost_serializers` to `external_bookings`
