+++ api/src/pcapi/core/external_bookings/boost/client.py
@@ -13,7 +13,6 @@ import pcapi.core.external_bookings.models as external_bookings_models
 import pcapi.core.providers.repository as providers_repository
 import pcapi.core.users.models as users_models
 from pcapi import settings
-from pcapi.connectors.serialization import boost_serializers
 from pcapi.core.external_bookings.decorators import catch_cinema_provider_request_timeout
 from pcapi.core.providers.models import BoostCinemaDetails
 from pcapi.utils import repository
@@ -23,14 +22,15 @@ from pcapi.utils.queue import add_to_queue
 
 from . import constants
 from . import exceptions
+from . import serializers
 
 
 logger = logging.getLogger(__name__)
 
 
 def get_pcu_pricing_if_exists(
-    showtime_pricing_list: list[boost_serializers.ShowtimePricing],
-) -> boost_serializers.ShowtimePricing | None:
+    showtime_pricing_list: list[serializers.ShowtimePricing],
+) -> serializers.ShowtimePricing | None:
     pcu_pricings = [
         pricing
         for pricing in showtime_pricing_list
@@ -176,7 +176,7 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
             )
 
         content = response.json()
-        login_info = parse_obj_as(boost_serializers.LoginBoost, content)
+        login_info = parse_obj_as(serializers.LoginBoost, content)
         token = login_info.token
         if not token:
             raise boost_exceptions.BoostLoginException("No token received from Boost API")
@@ -271,12 +271,12 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
         barcodes = list(set(barcodes))
         sale_cancel_items = []
         for barcode in barcodes:
-            sale_cancel_item = boost_serializers.SaleCancelItem(
+            sale_cancel_item = serializers.SaleCancelItem(
                 code=barcode, refundType=constants.BOOST_PASS_CULTURE_REFUND_TYPE
             )
             sale_cancel_items.append(sale_cancel_item)
 
-        sale_cancel = boost_serializers.SaleCancel(sales=sale_cancel_items)
+        sale_cancel = serializers.SaleCancel(sales=sale_cancel_items)
         self._authenticated_put(
             f"{self.cinema_details.cinemaUrl}api/sale/orderCancel",
             payload=sale_cancel.json(by_alias=True),
@@ -292,8 +292,8 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
         if not pcu_pricing:
             raise exceptions.BoostAPIException(f"pass Culture pricing not found for show {show_id}")
 
-        basket_items = [boost_serializers.BasketItem(idShowtimePricing=pcu_pricing.id, quantity=quantity)]
-        sale_body = boost_serializers.SaleRequest(
+        basket_items = [serializers.BasketItem(idShowtimePricing=pcu_pricing.id, quantity=quantity)]
+        sale_body = serializers.SaleRequest(
             codePayment=constants.BOOST_PASS_CULTURE_CODE_PAYMENT, basketItems=basket_items, idsBeforeSale=None
         )
 
@@ -302,7 +302,7 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
             f"{self.cinema_details.cinemaUrl}api/sale/complete",
             payload=sale_body.json(by_alias=True),
         )
-        sale_preparation = parse_obj_as(boost_serializers.SalePreparationResponse, sale_preparation_response)
+        sale_preparation = parse_obj_as(serializers.SalePreparationResponse, sale_preparation_response)
 
         # step 2: confirmation
         sale_body.idsBeforeSale = str(sale_preparation.data[0].id)
@@ -310,7 +310,7 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
             f"{self.cinema_details.cinemaUrl}api/sale/complete",
             payload=sale_body.json(by_alias=True),
         )
-        sale_confirmation_response = parse_obj_as(boost_serializers.SaleConfirmationResponse, sale_response)
+        sale_confirmation_response = parse_obj_as(serializers.SaleConfirmationResponse, sale_response)
         add_to_queue(
             bookings_constants.REDIS_EXTERNAL_BOOKINGS_NAME,
             {
@@ -364,7 +364,7 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
 
             data = self._authenticated_get(url, params=params)
 
-            collection = parse_obj_as(boost_serializers.ShowTimeCollection, data)
+            collection = parse_obj_as(serializers.ShowTimeCollection, data)
             items.extend(collection.data)
             total_pages = collection.totalPages
             next_page = collection.nextPage
@@ -374,9 +374,9 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
 
         return items
 
-    def get_showtime(self, showtime_id: int) -> boost_serializers.ShowTime4:
+    def get_showtime(self, showtime_id: int) -> serializers.ShowTime4:
         data = self._authenticated_get(f"{self.cinema_details.cinemaUrl}api/showtimes/{showtime_id}")
-        showtime_details = parse_obj_as(boost_serializers.ShowTimeDetails, data)
+        showtime_details = parse_obj_as(serializers.ShowTimeDetails, data)
         return showtime_details.data
 
     def get_movie_poster(self, image_url: str) -> bytes:
@@ -388,7 +388,7 @@ class BoostClientAPI(external_bookings_models.ExternalBookingsClientAPI):
 
         return api_response.content
 
-    def get_cinemas_attributs(self) -> list[boost_serializers.CinemaAttribut]:
+    def get_cinemas_attributs(self) -> list[serializers.CinemaAttribut]:
         data = self._authenticated_get(f"{self.cinema_details.cinemaUrl}api/cinemas/attributs")
-        attributs = parse_obj_as(boost_serializers.CinemaAttributCollection, data)
+        attributs = parse_obj_as(serializers.CinemaAttributCollection, data)
         return attributs.data
