+++ api/src/pcapi/core/external_bookings/boost/serializers.py
