+++ api/src/pcapi/local_providers/cinema_providers/boost/boost_stocks.py
@@ -8,8 +8,8 @@ import PIL
 
 from pcapi import settings
 from pcapi.connectors import thumb_storage
-from pcapi.connectors.serialization import boost_serializers
 from pcapi.core.categories import subcategories
+from pcapi.core.external_bookings.boost import serializers as boost_serializers
 from pcapi.core.external_bookings.boost.client import BoostClientAPI
 from pcapi.core.external_bookings.boost.client import get_pcu_pricing_if_exists
 from pcapi.core.offerers.models import Venue
